refactor(TanstackTable): seperate files into folders & refactor exports
